requirements.txt
customtkinter==5.0.3
requests==2.31.0
Pillow==9.5.0
openai==0.28.0